requirements.txt
django==2.0
djangorestframework==3.7.3

# Data Base
psycopg2==2.7.3

Pillow==4.3.0
